refactor: UI별 파일로 분리
