+++ src/index.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import TodoTable from './App';
+import TodoTable from './TodoTable';
+
 
 
 ReactDOM.render(
   <TodoTable />,
   document.getElementById('root')
 );
+
